+++ openeogeotrellis/config/config.py
@@ -74,6 +74,7 @@ class GpsBackendConfig(OpenEoBackendConfig):
 
     capabilities_backend_version: str = get_backend_version()
     capabilities_deploy_metadata: dict = attrs.Factory(_default_capabilities_deploy_metadata)
+    capabilities_extras: Optional[dict] = None
     processing_software = f"openeo-geopyspark-driver-{get_backend_version()}"
 
     oidc_providers: List[OidcProvider] = attrs.Factory(list)